# docs/swagger.yaml
definitions:
...
  util.CommonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /home/recommendSubject/create:
    post:
      consumes:
      - application/json
      description: 添加首页推荐专题
      parameters:
      - in: query
        name: id
        type: integer
      - in: query
        name: recommendStatus
        type: integer
      - in: query
        name: sort
        type: integer
      - in: query
        name: subjectId
        type: integer
      - in: query
        name: subjectName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 添加首页推荐专题
      tags:
      - 首页专题推荐管理
  /home/recommendSubject/delete:
    post:
      consumes:
      - application/json
      description: 批量删除推荐
      parameters:
      - description: ids
        in: query
        items:
          type: string
        name: ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 批量删除推荐
      tags:
      - 首页专题推荐管理
  /home/recommendSubject/list:
    get:
      consumes:
      - application/json
      description: 分页查询推荐
      parameters:
      - description: subjectName
        in: query
        name: subjectName
        type: string
      - description: recommendStatus
        in: query
        name: recommendStatus
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: string
      - description: 数量
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 分页查询推荐
      tags:
      - 首页专题推荐管理
  /home/recommendSubject/update/recommendStatus:
    post:
      consumes:
      - application/json
      description: 批量修改推荐状态
      parameters:
      - description: ids
        in: query
        items:
          type: string
        name: ids
        type: array
      - description: recommendStatus
        in: query
        name: recommendStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 批量修改推荐状态
      tags:
      - 首页专题推荐管理
  /home/recommendSubject/update/sort/{id}:
    post:
      consumes:
      - application/json
      description: 修改推荐排序
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改推荐排序
      tags:
      - 首页专题推荐管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"